Draft: split WorkingPlaneProxy from Draft.py
